+++ cash/src/main/java/ru/yandex/practicum/cash/dto/AccountsRequest.java
@@ -1,4 +1,6 @@
 package ru.yandex.practicum.cash.dto;
 
-public record AccountsRequest(Long userId, String currency, Double amount) {
+import java.math.BigDecimal;
+
+public record AccountsRequest(Long userId, String currency, BigDecimal amount) {
 }
