+++ cash/src/main/java/ru/yandex/practicum/cash/dto/NotificationRequest.java
@@ -1,4 +1,6 @@
 package ru.yandex.practicum.cash.dto;
 
-public record NotificationRequest(Long userId, String type, String message, Double amount, String currency) {
+import java.math.BigDecimal;
+
+public record NotificationRequest(Long userId, String type, String message, BigDecimal amount, String currency) {
 }
